utilities/requirements.txt
# Python dependencies for mousebot serial utilities
# Install with: pip3 install -r requirements.txt

pyserial>=3.5
protobuf>=3.20.0
